Moved CafeTableModel into newly created model package
