+++ f21as_coursework/src/model/package-info.java
@@ -1,0 +1,1 @@
+package model;
